pyproject.toml
[project]
name = "notes-semantic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
